Add files via upload
This file runs, but the accuracy is basically chance which is weird. Renato is still trying to figure this out(?)

We can extend this to the 3 by 3 case by requiring the line of three shapes to all be the same or all be different (e.g. cirlce, triangle or square). We can add more attributes such as introducing patterns or different numbers of shapes, and requiring them to all be different or all be the same.